--- src/App.css
@@ -1,4 +1,0 @@
-.app-container {
-  width: 400px;
-  margin: 100px auto;
-}
