+++ src/components/Minutes.tsx
@@ -1,17 +1,17 @@
 import React from 'react'
 import { Radio, Space, Select, InputNumber, Row, Col } from 'antd'
 import type { RadioChangeEvent } from 'antd'
-import styles from '../ReactCron.less'
+import styles from '../styles/index.less'
 import { useTranslation } from 'react-i18next'
 
-export type HoursProps = {
+export type MinutesProps = {
   value: any
   onChange?: (value: number[] | number, filedKey: string) => void
 }
 
-const Hours: React.FC<HoursProps> = (props) => {
+const Minutes: React.FC<MinutesProps> = (props) => {
   const { t } = useTranslation()
-  const HoursOptions = new Array(24).fill(0).map((value, index) => ({
+  const MinutesOptions = new Array(60).fill(0).map((value, index) => ({
     value: index
   }))
 
@@ -27,7 +27,7 @@ const Hours: React.FC<HoursProps> = (props) => {
     }
   }
 
-  const onChangeHour = (value: number[] | number) => {
+  const onChangeMinute = (value: number[] | number) => {
     if (props.onChange) {
       props.onChange(value, 'specificSpecific')
     }
@@ -37,14 +37,14 @@ const Hours: React.FC<HoursProps> = (props) => {
     <div className={styles.tabContent}>
       <Radio.Group onChange={onChangeType} value={props.value.cronEvery}>
         <Space direction="vertical">
-          <Radio value={1}>{t('Hours.every')}</Radio>
+          <Radio value={1}>{t('Minutes.every')}</Radio>
           <Row align="middle">
             <Col>
               <Radio value={2}></Radio>
             </Col>
             <Col>
               <span className={styles.optionLabel}>
-                {t('Hours.interval.0')}
+                {t('Minutes.interval.0')}
               </span>
               <InputNumber
                 value={props.value.incrementIncrement}
@@ -53,7 +53,7 @@ const Hours: React.FC<HoursProps> = (props) => {
                 onChange={(value) => onChangeInput(value, 'incrementIncrement')}
               />
               <span className={styles.optionLabel}>
-                {t('Hours.interval.1')}
+                {t('Minutes.interval.1')}
               </span>
               <InputNumber
                 value={props.value.incrementStart}
@@ -62,7 +62,7 @@ const Hours: React.FC<HoursProps> = (props) => {
                 onChange={(value) => onChangeInput(value, 'incrementStart')}
               />
               <span className={styles.optionLabel}>
-                {t('Hours.interval.2')}
+                {t('Minutes.interval.2')}
               </span>
             </Col>
           </Row>
@@ -71,14 +71,16 @@ const Hours: React.FC<HoursProps> = (props) => {
               <Radio value={3}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Hours.specific')}</span>
+              <span className={styles.optionLabel}>
+                {t('Minutes.specific')}
+              </span>
               <Select
                 mode="multiple"
                 allowClear
                 style={{ width: 280 }}
                 value={props.value.specificSpecific}
-                options={HoursOptions}
-                onChange={onChangeHour}
+                options={MinutesOptions}
+                onChange={onChangeMinute}
               />
             </Col>
           </Row>
@@ -87,21 +89,21 @@ const Hours: React.FC<HoursProps> = (props) => {
               <Radio value={4}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Hours.cycle.0')}</span>
+              <span className={styles.optionLabel}>{t('Minutes.cycle.0')}</span>
               <InputNumber
                 value={props.value.rangeStart}
                 min={1}
                 max={60}
                 onChange={(value) => onChangeInput(value, 'rangeStart')}
               />
-              <span className={styles.optionLabel}>{t('Hours.cycle.1')}</span>
+              <span className={styles.optionLabel}>{t('Minutes.cycle.1')}</span>
               <InputNumber
                 value={props.value.rangeEnd}
                 min={0}
                 max={59}
                 onChange={(value) => onChangeInput(value, 'rangeEnd')}
               />
-              <span className={styles.optionLabel}>{t('Hours.cycle.2')}</span>
+              <span className={styles.optionLabel}>{t('Minutes.cycle.2')}</span>
             </Col>
           </Row>
         </Space>
@@ -110,4 +112,4 @@ const Hours: React.FC<HoursProps> = (props) => {
   )
 }
 
-export default Hours
+export default Minutes
