+++ src/components/Month.tsx
@@ -1,18 +1,18 @@
 import React from 'react'
 import { Radio, Space, Select, InputNumber, Row, Col } from 'antd'
 import type { RadioChangeEvent } from 'antd'
-import styles from '../ReactCron.less'
+import styles from '../styles/index.less'
 import { useTranslation } from 'react-i18next'
 
-export type SecondsProps = {
+export type MonthProps = {
   value: any
   onChange?: (value: number[] | number, filedKey: string) => void
 }
 
-const Seconds: React.FC<SecondsProps> = (props) => {
+const Month: React.FC<MonthProps> = (props) => {
   const { t } = useTranslation()
-  const secondsOptions = new Array(60).fill(0).map((value, index) => ({
-    value: index
+  const monthOptions = new Array(12).fill(0).map((value, index) => ({
+    value: index + 1
   }))
 
   const onChangeType = (e: RadioChangeEvent) => {
@@ -37,32 +37,32 @@ const Seconds: React.FC<SecondsProps> = (props) => {
     <div className={styles.tabContent}>
       <Radio.Group onChange={onChangeType} value={props.value.cronEvery}>
         <Space direction="vertical">
-          <Radio value={1}>{t('Seconds.every')}</Radio>
+          <Radio value={1}>{t('Month.every')}</Radio>
           <Row align="middle">
             <Col>
               <Radio value={2}></Radio>
             </Col>
             <Col>
               <span className={styles.optionLabel}>
-                {t('Seconds.interval.0')}
+                {t('Month.interval.0')}
               </span>
               <InputNumber
                 value={props.value.incrementIncrement}
-                min={1}
-                max={60}
+                min={0}
+                max={12}
                 onChange={(value) => onChangeInput(value, 'incrementIncrement')}
               />
               <span className={styles.optionLabel}>
-                {t('Seconds.interval.1')}
+                {t('Month.interval.1')}
               </span>
               <InputNumber
                 value={props.value.incrementStart}
                 min={0}
-                max={59}
+                max={12}
                 onChange={(value) => onChangeInput(value, 'incrementStart')}
               />
               <span className={styles.optionLabel}>
-                {t('Seconds.interval.2') || ''}
+                {t('Month.interval.2')}
               </span>
             </Col>
           </Row>
@@ -71,15 +71,13 @@ const Seconds: React.FC<SecondsProps> = (props) => {
               <Radio value={3}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>
-                {t('Seconds.specific')}
-              </span>
+              <span className={styles.optionLabel}>{t('Month.specific')}</span>
               <Select
                 mode="multiple"
                 allowClear
-                style={{ width: 280 }}
+                style={{ width: 260 }}
                 value={props.value.specificSpecific}
-                options={secondsOptions}
+                options={monthOptions}
                 onChange={onChangeSecond}
               />
             </Col>
@@ -89,21 +87,21 @@ const Seconds: React.FC<SecondsProps> = (props) => {
               <Radio value={4}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Seconds.cycle.0')}</span>
+              <span className={styles.optionLabel}>{t('Month.cycle.0')}</span>
               <InputNumber
                 value={props.value.rangeStart}
                 min={1}
-                max={60}
+                max={12}
                 onChange={(value) => onChangeInput(value, 'rangeStart')}
               />
-              <span className={styles.optionLabel}>{t('Seconds.cycle.1')}</span>
+              <span className={styles.optionLabel}>{t('Month.cycle.1')}</span>
               <InputNumber
                 value={props.value.rangeEnd}
-                min={0}
-                max={59}
+                min={1}
+                max={12}
                 onChange={(value) => onChangeInput(value, 'rangeEnd')}
               />
-              <span className={styles.optionLabel}>{t('Seconds.cycle.2')}</span>
+              <span className={styles.optionLabel}>{t('Month.cycle.2')}</span>
             </Col>
           </Row>
         </Space>
@@ -112,4 +110,4 @@ const Seconds: React.FC<SecondsProps> = (props) => {
   )
 }
 
-export default Seconds
+export default Month
