+++ src/components/Seconds.tsx
@@ -1,23 +1,18 @@
 import React from 'react'
-import moment from 'moment'
 import { Radio, Space, Select, InputNumber, Row, Col } from 'antd'
 import type { RadioChangeEvent } from 'antd'
-import styles from '../ReactCron.less'
+import styles from '../styles/index.less'
 import { useTranslation } from 'react-i18next'
 
-const maxRange = 20
-const currentYear = moment().year()
-const maxRangeYear = moment().add(maxRange, 'year').year()
-
-export type YearProps = {
+export type SecondsProps = {
   value: any
   onChange?: (value: number[] | number, filedKey: string) => void
 }
 
-const Year: React.FC<YearProps> = (props) => {
+const Seconds: React.FC<SecondsProps> = (props) => {
   const { t } = useTranslation()
-  const yearOptions = new Array(maxRange).fill(0).map((value, index) => ({
-    value: moment().add(index, 'year').year()
+  const secondsOptions = new Array(60).fill(0).map((value, index) => ({
+    value: index
   }))
 
   const onChangeType = (e: RadioChangeEvent) => {
@@ -42,32 +37,32 @@ const Year: React.FC<YearProps> = (props) => {
     <div className={styles.tabContent}>
       <Radio.Group onChange={onChangeType} value={props.value.cronEvery}>
         <Space direction="vertical">
-          <Radio value={1}>{t('Year.every')}</Radio>
+          <Radio value={1}>{t('Seconds.every')}</Radio>
           <Row align="middle">
             <Col>
               <Radio value={2}></Radio>
             </Col>
             <Col>
               <span className={styles.optionLabel}>
-                {t('Year.interval.0')}
+                {t('Seconds.interval.0')}
               </span>
               <InputNumber
                 value={props.value.incrementIncrement}
                 min={1}
-                max={maxRange}
+                max={60}
                 onChange={(value) => onChangeInput(value, 'incrementIncrement')}
               />
               <span className={styles.optionLabel}>
-                {t('Year.interval.1')}
+                {t('Seconds.interval.1')}
               </span>
               <InputNumber
                 value={props.value.incrementStart}
-                min={currentYear}
-                max={maxRangeYear}
+                min={0}
+                max={59}
                 onChange={(value) => onChangeInput(value, 'incrementStart')}
               />
               <span className={styles.optionLabel}>
-                {t('Year.interval.2')}
+                {t('Seconds.interval.2') || ''}
               </span>
             </Col>
           </Row>
@@ -76,13 +71,15 @@ const Year: React.FC<YearProps> = (props) => {
               <Radio value={3}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Year.specific')}</span>
+              <span className={styles.optionLabel}>
+                {t('Seconds.specific')}
+              </span>
               <Select
                 mode="multiple"
                 allowClear
-                style={{ width: 260 }}
+                style={{ width: 280 }}
                 value={props.value.specificSpecific}
-                options={yearOptions}
+                options={secondsOptions}
                 onChange={onChangeSecond}
               />
             </Col>
@@ -92,21 +89,21 @@ const Year: React.FC<YearProps> = (props) => {
               <Radio value={4}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Year.cycle.0')}</span>
+              <span className={styles.optionLabel}>{t('Seconds.cycle.0')}</span>
               <InputNumber
                 value={props.value.rangeStart}
-                min={currentYear}
-                max={maxRangeYear}
+                min={1}
+                max={60}
                 onChange={(value) => onChangeInput(value, 'rangeStart')}
               />
-              <span className={styles.optionLabel}>{t('Year.cycle.1')}</span>
+              <span className={styles.optionLabel}>{t('Seconds.cycle.1')}</span>
               <InputNumber
                 value={props.value.rangeEnd}
-                min={currentYear}
-                max={maxRangeYear}
+                min={0}
+                max={59}
                 onChange={(value) => onChangeInput(value, 'rangeEnd')}
               />
-              <span className={styles.optionLabel}>{t('Year.cycle.2')}</span>
+              <span className={styles.optionLabel}>{t('Seconds.cycle.2')}</span>
             </Col>
           </Row>
         </Space>
@@ -115,4 +112,4 @@ const Year: React.FC<YearProps> = (props) => {
   )
 }
 
-export default Year
+export default Seconds
