+++ src/components/Year.tsx
@@ -1,18 +1,23 @@
 import React from 'react'
+import moment from 'moment'
 import { Radio, Space, Select, InputNumber, Row, Col } from 'antd'
 import type { RadioChangeEvent } from 'antd'
-import styles from '../ReactCron.less'
+import styles from '../styles/index.less'
 import { useTranslation } from 'react-i18next'
 
-export type MonthProps = {
+const maxRange = 20
+const currentYear = moment().year()
+const maxRangeYear = moment().add(maxRange, 'year').year()
+
+export type YearProps = {
   value: any
   onChange?: (value: number[] | number, filedKey: string) => void
 }
 
-const Month: React.FC<MonthProps> = (props) => {
+const Year: React.FC<YearProps> = (props) => {
   const { t } = useTranslation()
-  const monthOptions = new Array(12).fill(0).map((value, index) => ({
-    value: index + 1
+  const yearOptions = new Array(maxRange).fill(0).map((value, index) => ({
+    value: moment().add(index, 'year').year()
   }))
 
   const onChangeType = (e: RadioChangeEvent) => {
@@ -37,33 +42,27 @@ const Month: React.FC<MonthProps> = (props) => {
     <div className={styles.tabContent}>
       <Radio.Group onChange={onChangeType} value={props.value.cronEvery}>
         <Space direction="vertical">
-          <Radio value={1}>{t('Month.every')}</Radio>
+          <Radio value={1}>{t('Year.every')}</Radio>
           <Row align="middle">
             <Col>
               <Radio value={2}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>
-                {t('Month.interval.0')}
-              </span>
+              <span className={styles.optionLabel}>{t('Year.interval.0')}</span>
               <InputNumber
                 value={props.value.incrementIncrement}
-                min={0}
-                max={12}
+                min={1}
+                max={maxRange}
                 onChange={(value) => onChangeInput(value, 'incrementIncrement')}
               />
-              <span className={styles.optionLabel}>
-                {t('Month.interval.1')}
-              </span>
+              <span className={styles.optionLabel}>{t('Year.interval.1')}</span>
               <InputNumber
                 value={props.value.incrementStart}
-                min={0}
-                max={12}
+                min={currentYear}
+                max={maxRangeYear}
                 onChange={(value) => onChangeInput(value, 'incrementStart')}
               />
-              <span className={styles.optionLabel}>
-                {t('Month.interval.2')}
-              </span>
+              <span className={styles.optionLabel}>{t('Year.interval.2')}</span>
             </Col>
           </Row>
           <Row align="middle">
@@ -71,13 +70,13 @@ const Month: React.FC<MonthProps> = (props) => {
               <Radio value={3}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Month.specific')}</span>
+              <span className={styles.optionLabel}>{t('Year.specific')}</span>
               <Select
                 mode="multiple"
                 allowClear
                 style={{ width: 260 }}
                 value={props.value.specificSpecific}
-                options={monthOptions}
+                options={yearOptions}
                 onChange={onChangeSecond}
               />
             </Col>
@@ -87,21 +86,21 @@ const Month: React.FC<MonthProps> = (props) => {
               <Radio value={4}></Radio>
             </Col>
             <Col>
-              <span className={styles.optionLabel}>{t('Month.cycle.0')}</span>
+              <span className={styles.optionLabel}>{t('Year.cycle.0')}</span>
               <InputNumber
                 value={props.value.rangeStart}
-                min={1}
-                max={12}
+                min={currentYear}
+                max={maxRangeYear}
                 onChange={(value) => onChangeInput(value, 'rangeStart')}
               />
-              <span className={styles.optionLabel}>{t('Month.cycle.1')}</span>
+              <span className={styles.optionLabel}>{t('Year.cycle.1')}</span>
               <InputNumber
                 value={props.value.rangeEnd}
-                min={1}
-                max={12}
+                min={currentYear}
+                max={maxRangeYear}
                 onChange={(value) => onChangeInput(value, 'rangeEnd')}
               />
-              <span className={styles.optionLabel}>{t('Month.cycle.2')}</span>
+              <span className={styles.optionLabel}>{t('Year.cycle.2')}</span>
             </Col>
           </Row>
         </Space>
@@ -110,4 +109,4 @@ const Month: React.FC<MonthProps> = (props) => {
   )
 }
 
-export default Month
+export default Year
