+++ src/language/i18n.ts
