+++ src/language/resources.ts
