+++ src/styles/index.less
@@ -1,0 +1,33 @@
+.cronContainer {
+    max-width: 600px;
+
+    .result {
+        font-size: 16px;
+        text-align: center;
+        margin: 8px;
+    }
+
+    .tabContent {
+        :global {
+            .ant-col {
+                line-height: 30px;
+                vertical-align: middle;
+            }
+        }
+
+        .optionLabel {
+            display: inline-block;
+            margin-right: 0.5em;
+        }
+
+        div+.optionLabel {
+            margin-left: 0.5em;
+        }
+    }
+
+    .control {
+        width: 100%;
+        margin-top: 10px;
+        text-align: center;
+    }
+}
