+++ src/test/App.css
@@ -1,0 +1,4 @@
+.app-container {
+    width: 400px;
+    margin: 100px auto;
+  }
